+++ services/scheduler/service.go
@@ -52,7 +52,10 @@ type Service interface {
 	ScheduleJob(ctx context.Context, name string, runtime time.Time, job JobFunc, data interface{}) error
 
 	// SchedulePeriodicJob schedules a job to run in a loop.
-	SchedulePeriodicJob(ctx context.Context, name string, runtime RuntimeFunc, runtineData interface{}, job JobFunc, jobData interface{}) error
+	// The loop starts by calling runtimeFunc, which sets the time for the first run.
+	// Once the time as specified by runtimeFunc is met, jobFunc is called.
+	// Once jobFunc returns, go back to the beginning of the loop.
+	SchedulePeriodicJob(ctx context.Context, name string, runtime RuntimeFunc, runtimeData interface{}, job JobFunc, jobData interface{}) error
 
 	// CancelJob cancels a known job.
 	// If this is a period job then all future instances are cancelled.