+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+Development
+  - tidy up trace logging for scheduler
+
 0.6.2
   - do not attempt to aggregate a failed attestation
   - error appropriately when misconfigured